# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Drust Daam APIs",
        "description": "User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
          "url": "https://drust-daam.herokuapp.com/api/",
          "description": "Live server"
        },
        {
          "url": "http://localhost:5000/api",
          "description": "Local server"
        }
      ],
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/regulate/item": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Regulate" ],
                "summary": "Get all regulare items in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regulate"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": [
                    "Regulate"
                ],
                "summary": "Create a new regulate items in system",
                "requestBody": {
                    "description": "Regulate Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Regulate"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
        "/regulate/item/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the regulate items that we want to match",
                    "type": "string"
                }
            ],
            "get": {
                "security": [{ "bearerAuth": [] }],
                "summary": "get regulate item with itemId",
                "tags": [
                    "Regulate"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "get regulate item with itemId",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regulate"
                        }
                    },
                    "404": {
                        "description": "Failed regulate items not found."
                    }
                }
            }
        },
        "/regulate/itembydate/{date}": {
            
            "get": {
                "security": [{ "bearerAuth": [] }],
                "summary": "get item with date",
                "tags": [
                    "Regulate"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "description": "get item with date",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Regulate"
                        }
                    },
                    "404": {
                        "description": "Failed regulate items not found."
                    }
                }
            }
        },
        "/unit": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Unit" ],
                "summary": "Get all unit in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Unit"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": ["Unit"],
                "summary": "Create a new unit in system",
                "requestBody": {
                    "description": "Unit Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Unit"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
        "/unit/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the unit that we want to match",
                    "type": "string"
                }
            ],
            "delete": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Delete unit with given ID",
                "tags": [
                    "Unit"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete unit with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Unit"
                        }
                    },
                    "404": {
                        "description": "Failed unit not found."
                    }
                }
            },
            "put": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Update unit with given ID",
                "tags": [
                    "Unit"
                ],
                "requestBody": {
                    "description": "Unit Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Unit"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "unit with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Unit"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    },
                    "404": {
                        "description": "Failed unit not found."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "Regulate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "itemId": {
                    "type": "string"
                },
                "minPrice": {
                    "type": "number"
                },
                "maxPrice": {
                    "type": "number"
                }
            }
        },
...
        "Unit": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                }
            }
        },
...
    }
}